+++ client/rest-high-level/src/test/java/org/opensearch/client/AbstractResponseTestCase.java
@@ -32,7 +32,6 @@
 package org.opensearch.client;
 
 import org.opensearch.common.bytes.BytesReference;
-import org.opensearch.common.collect.ImmutableOpenMap;
 import org.opensearch.common.xcontent.LoggingDeprecationHandler;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.ToXContent;
@@ -43,10 +42,6 @@ import org.opensearch.common.xcontent.XContentType;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.io.IOException;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Set;
 
 /**
  * Base class for HLRC response parsing tests.
@@ -102,18 +97,5 @@ public abstract class AbstractResponseTestCase<S extends ToXContent, C> extends 
      */
     protected ToXContent.Params getParams() {
         return ToXContent.EMPTY_PARAMS;
-    }
-
-    protected static <T> void assertMapEquals(ImmutableOpenMap<String, T> expected, Map<String, T> actual) {
-        Set<String> expectedKeys = new HashSet<>();
-        Iterator<String> keysIt = expected.keysIt();
-        while (keysIt.hasNext()) {
-            expectedKeys.add(keysIt.next());
-        }
-
-        assertEquals(expectedKeys, actual.keySet());
-        for (String key : expectedKeys) {
-            assertEquals(expected.get(key), actual.get(key));
-        }
     }
 }