+++ server/src/main/java/org/opensearch/action/admin/indices/get/GetIndexResponse.java
@@ -32,14 +32,12 @@
 
 package org.opensearch.action.admin.indices.get;
 
-import com.carrotsearch.hppc.cursors.ObjectObjectCursor;
 import org.opensearch.LegacyESVersion;
 import org.opensearch.Version;
 import org.opensearch.action.ActionResponse;
 import org.opensearch.cluster.metadata.AliasMetadata;
 import org.opensearch.cluster.metadata.MappingMetadata;
 import org.opensearch.common.Strings;
-import org.opensearch.common.collect.ImmutableOpenMap;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.settings.Settings;
@@ -66,18 +64,18 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
 
     private Map<String, MappingMetadata> mappings = Map.of();
     private Map<String, List<AliasMetadata>> aliases = Map.of();
-    private ImmutableOpenMap<String, Settings> settings = ImmutableOpenMap.of();
-    private ImmutableOpenMap<String, Settings> defaultSettings = ImmutableOpenMap.of();
-    private ImmutableOpenMap<String, String> dataStreams = ImmutableOpenMap.of();
+    private Map<String, Settings> settings = Map.of();
+    private Map<String, Settings> defaultSettings = Map.of();
+    private Map<String, String> dataStreams = Map.of();
     private final String[] indices;
 
     public GetIndexResponse(
         String[] indices,
         Map<String, MappingMetadata> mappings,
         final Map<String, List<AliasMetadata>> aliases,
-        ImmutableOpenMap<String, Settings> settings,
-        ImmutableOpenMap<String, Settings> defaultSettings,
-        ImmutableOpenMap<String, String> dataStreams
+        final Map<String, Settings> settings,
+        final Map<String, Settings> defaultSettings,
+        final Map<String, String> dataStreams
     ) {
         this.indices = indices;
         // to have deterministic order
@@ -89,13 +87,13 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
             this.aliases = Collections.unmodifiableMap(aliases);
         }
         if (settings != null) {
-            this.settings = settings;
+            this.settings = Collections.unmodifiableMap(settings);
         }
         if (defaultSettings != null) {
-            this.defaultSettings = defaultSettings;
+            this.defaultSettings = Collections.unmodifiableMap(defaultSettings);
         }
         if (dataStreams != null) {
-            this.dataStreams = dataStreams;
+            this.dataStreams = Collections.unmodifiableMap(dataStreams);
         }
     }
 
@@ -141,27 +139,27 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
         aliases = Collections.unmodifiableMap(aliasesMapBuilder);
 
         int settingsSize = in.readVInt();
-        ImmutableOpenMap.Builder<String, Settings> settingsMapBuilder = ImmutableOpenMap.builder();
+        final Map<String, Settings> settingsMapBuilder = new HashMap<>();
         for (int i = 0; i < settingsSize; i++) {
             String key = in.readString();
             settingsMapBuilder.put(key, Settings.readSettingsFromStream(in));
         }
-        settings = settingsMapBuilder.build();
+        settings = Collections.unmodifiableMap(settingsMapBuilder);
 
-        ImmutableOpenMap.Builder<String, Settings> defaultSettingsMapBuilder = ImmutableOpenMap.builder();
+        final Map<String, Settings> defaultSettingsMapBuilder = new HashMap<>();
         int defaultSettingsSize = in.readVInt();
         for (int i = 0; i < defaultSettingsSize; i++) {
             defaultSettingsMapBuilder.put(in.readString(), Settings.readSettingsFromStream(in));
         }
-        defaultSettings = defaultSettingsMapBuilder.build();
+        defaultSettings = Collections.unmodifiableMap(defaultSettingsMapBuilder);
 
         if (in.getVersion().onOrAfter(LegacyESVersion.V_7_8_0)) {
-            ImmutableOpenMap.Builder<String, String> dataStreamsMapBuilder = ImmutableOpenMap.builder();
+            final Map<String, String> dataStreamsMapBuilder = new HashMap<>();
             int dataStreamsSize = in.readVInt();
             for (int i = 0; i < dataStreamsSize; i++) {
                 dataStreamsMapBuilder.put(in.readString(), in.readOptionalString());
             }
-            dataStreams = dataStreamsMapBuilder.build();
+            dataStreams = Collections.unmodifiableMap(dataStreamsMapBuilder);
         }
     }
 
@@ -189,15 +187,15 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
         return aliases();
     }
 
-    public ImmutableOpenMap<String, Settings> settings() {
+    public Map<String, Settings> settings() {
         return settings;
     }
 
-    public ImmutableOpenMap<String, String> dataStreams() {
+    public Map<String, String> dataStreams() {
         return dataStreams;
     }
 
-    public ImmutableOpenMap<String, String> getDataStreams() {
+    public Map<String, String> getDataStreams() {
         return dataStreams();
     }
 
@@ -209,11 +207,11 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
      * via {@link #settings()}.
      * See also {@link GetIndexRequest#includeDefaults(boolean)}
      */
-    public ImmutableOpenMap<String, Settings> defaultSettings() {
+    public Map<String, Settings> defaultSettings() {
         return defaultSettings;
     }
 
-    public ImmutableOpenMap<String, Settings> getSettings() {
+    public Map<String, Settings> getSettings() {
         return settings();
     }
 
@@ -266,20 +264,20 @@ public class GetIndexResponse extends ActionResponse implements ToXContentObject
             }
         }
         out.writeVInt(settings.size());
-        for (ObjectObjectCursor<String, Settings> indexEntry : settings) {
-            out.writeString(indexEntry.key);
-            Settings.writeSettingsToStream(indexEntry.value, out);
+        for (final Map.Entry<String, Settings> indexEntry : settings.entrySet()) {
+            out.writeString(indexEntry.getKey());
+            Settings.writeSettingsToStream(indexEntry.getValue(), out);
         }
         out.writeVInt(defaultSettings.size());
-        for (ObjectObjectCursor<String, Settings> indexEntry : defaultSettings) {
-            out.writeString(indexEntry.key);
-            Settings.writeSettingsToStream(indexEntry.value, out);
+        for (final Map.Entry<String, Settings> indexEntry : defaultSettings.entrySet()) {
+            out.writeString(indexEntry.getKey());
+            Settings.writeSettingsToStream(indexEntry.getValue(), out);
         }
         if (out.getVersion().onOrAfter(LegacyESVersion.V_7_8_0)) {
             out.writeVInt(dataStreams.size());
-            for (ObjectObjectCursor<String, String> indexEntry : dataStreams) {
-                out.writeString(indexEntry.key);
-                out.writeOptionalString(indexEntry.value);
+            for (final Map.Entry<String, String> indexEntry : dataStreams.entrySet()) {
+                out.writeString(indexEntry.getKey());
+                out.writeOptionalString(indexEntry.getValue());
             }
         }
     }